requirements.txt
Django==5.2
gunicorn==23.0.0
pillow==11.1.0
requests==2.32.3
environs==14.1.1
psycopg2-binary==2.9.10
django-cors-headers==4.7.0
djangorestframework==3.16.0